Management of remote desktop connection requests.
This program receives all requests to log in to Windows servers by remote desktop and makes it easier to control the connection to your server.

You can see the login requests in your Telegram bot or our suggested bot.

It is possible that your Windows Defender server will identify this file as a virus because this program will request the registration of startup in the registry if you approve it.

To get help, type the help word into the program

The compiled program is matched with the proposed Telegram bot, and you can use the set bot to control inputs without creating a new bot in Telegram.
When starting the program, a settings.json file is created next to the program, which you can edit as you like (following the rules of the json file).
Finally, just start the @RDManagerBot bot in Telegram or introduce your bot to the program and set the Telegram ID code in the program. From now on, any request to enter the server will be sent to your bot in Telegram.